+++ src/keypin/core.clj
@@ -306,6 +306,8 @@
   |`:source`    |Source or container (of reference type, e.g. atom/agent/promise etc.) of key/value pairs     |
   |`:pre-xform` |Middleware function `(fn [option-map]) -> option-map` used before key definition is created  |
   |`:post-xform`|Middleware function `(fn [keypin.type.KeyAttributes]) -> keypin.type.KeyAttributes`          |
+  |`:cmsym-meta`|Metadata to apply to `config-map` argument symbol in the function definition, default none   |
+  |`:nfsym-meta`|Metadata to apply to `not-found` argument symbol in the function definition, default none    |
 
   The `validator` is a predicate `(fn [parsed-value]) -> boolean` that returns `true` for valid values,
   `false` otherwise.
@@ -363,9 +365,12 @@
                 (map (fn [[each-sym [the-key validator descrip options :as each-vec]]]
                        (let [;; descrip (nth each-vec 2 "No description")
                              ;; options (nth each-vec 3 {})  ; 4th element is option map
+                             cm-sym  (with-meta 'config-map (:cmsym-meta options))
+                             nf-sym  (with-meta 'not-found (:nfsym-meta options))
                              arities (if (:source options)
-                                       ''([] [config-map] [config-map not-found])
-                                       ''([config-map] [config-map not-found]))
+                                       `'([] [~cm-sym] [~cm-sym ~nf-sym]) ; ''([] [config-map] [config-map not-found])
+                                       `'([~cm-sym] [~cm-sym ~nf-sym])    ; ''([config-map] [config-map not-found])
+                                       )
                              meta-fn (partial merge {:arglists arities :doc descrip})
                              def-sym (vary-meta each-sym meta-fn)
                              arg-map (conj (get each-vec 3 {})