+++ CHANGES.md
@@ -20,8 +20,11 @@
 
 ## [WIP] 0.8.2 / 2021-February-??
 
-- [Todo] Default metadata in `keypin.core/*vary-defkey-meta*` for use by `keypin.core/defkey`
+- Add `keypin.core/defkey` option kwargs to attach metadata to function args
+  - `:cmsym-meta` for argument `config-map`
+  - `:nfsym-meta` for argument `not-found`
 - Add predicate `network-port?` for network port 0 through 65535
+- [Todo] Docstring formatting for Cljdoc
 
 
 ## 0.8.1 / 2020-August-10